requirements.txt
langchain
streamlit
streamlit_chat
OpenAI
python-dotenv
tensorflow==2.12.*
transformers